refactor: split widget.js into focused component modules
Extract form handling, select mode, items panel, messages, and keyboard
shortcuts into separate modules under assets/component/. The widget.js
entry point is now a slim orchestrator that creates a shared context
object and wires up the components.

Closes #70
